Move cleaning function into global util
