+++ source/views/student-orgs/detail.android.js
@@ -5,11 +5,11 @@ import moment from 'moment'
 import {Card} from '../components/card'
 import * as c from '../components/colors'
 import type {StudentOrgType} from './types'
-import type {ContactPersonType} from './types'
 import type {TopLevelViewPropsType} from '../types'
 import Communications from 'react-native-communications'
 import openUrl from '../components/open-url'
-import cleanOrg from './clean-org'
+import {cleanOrg} from './util'
+import {showNameOrEmail} from './util'
 
 const styles = StyleSheet.create({
   name: {
@@ -67,14 +67,6 @@ export class StudentOrgsDetailView extends React.Component {
     navigation: {state: {params: {org: StudentOrgType}}},
   }
 
-  showNameOrEmail = ({c}: {c: ContactPersonType}) => {
-    if (!c.firstName.trim() && !c.lastName.trim()) {
-      return `${c.email}`
-    }
-
-    return `${c.firstName} ${c.lastName} (${c.email})`
-  }
-
   // Using Communications because `mailTo` complains about
   // the lack of an available Activity...
   openEmail = (email: string, org: string) => {
@@ -127,7 +119,7 @@ export class StudentOrgsDetailView extends React.Component {
                   onPress={() => this.openEmail(c.email, orgName)}
                 >
                   {c.title ? c.title + ': ' : ''}
-                  {this.showNameOrEmail({c})}
+                  {showNameOrEmail(c)}
                 </Text>,
               )}
             </Card>
